[wasm] Split the token half of SignatureMapper into its own file
SignatureMapper mixes two things: reflection over scanned assemblies, which
needs a LogAdapter and so drags in Microsoft.Build, and a pure mapping from
signature tokens to native types, which needs nothing. Make it partial and move
the pure half out, so a test can compile it directly instead of pulling MSBuild
into a compiler test assembly.

No behaviour change: the generator emits a byte-identical portable entrypoint
table and an identical interp-to-managed table afterwards.
